Add files via upload
Catalina implements Tomcat's actual servlet specification; when we start up our Tomcat server, it will be faster and more convenient for us to launch the server by using Catalina.

We figure as local attackers who have access to the server as tomcat users (for example, through a vulnerability in a web application) and change the log file with a symlink to any system file. Then we can escalate our privileges to root once Tomcat init script (running as root) re-opens the catalina.out file after a service restart, reboot etc.

Additionally,  we have already had tomcat accounts at the time of exploitation, so we could also kill the tomcat processes to meet the need for a restart.